Add files via upload
Objective.
Bearing is an important component in contemporary mechanical equipment, its main function is to support the mechanical rotating body, reduce the friction coefficient during its movement, and ensure its rotary accuracy, so that the quality and reliability of the bearing has extremely demanding requirements.
At present, in order to screen out the bearings that meet the requirements, the manual visual inspection method is generally used in industrial production. This is a high cost and inefficient method of manual training. To address this pain point, we combine the current rapid development of image processing, pattern recognition, deep learning and other related science to solve the interdisciplinary problem of "how to apply computer vision to industrial product inspection", and then form a set of "vision" link The feedback system can be used to improve the productivity and quality of industrial production by using the image inspection technology which is widely used in the automatic inspection of industrial products.
Basic idea.
We designed a hardware and software system that can be used in the detection of bearing surface defects, through which the bearing surface defects can be accurately and automatically screened. The main steps are as follows: first (image acquisition), under the condition of coaxial illumination, the coaxial light image of the target bearing is acquired; under the condition of multi-angle and multi-light source illumination, the multi-light source image of the target bearing is acquired. Secondly (bearing localization), receive and process the coaxial optical image in the image acquisition module with the bearing center coordinates. Again (bearing region segmentation), reads the bearing center coordinates and divides the bearing into three different regions. Then (character recognition), reads the segmented center ring region of the image and uses the neural network to identify the trademark features. Finally (defect detection), receives the multi-angle light source image. And four kinds of defects (gap, stain, flat cover and scratch) are identified.
The main processing flow is shown in Figure 1.